Separate files agree with classes and create 'src' folder
