+++ src/capitalize.rb
@@ -1,0 +1,7 @@
+require_relative './decorator'
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @name.correct_name.capitalize
+  end
+end
